[Perpetual] move SLIPPAGE_DEFAULT to fix init order, fix minor issues
